\name{hypot}
\alias{hypot}
\title{sqrt(a^2 + b^2) without under/overflow}
\description{\code{hypot} sqrt(a^2 + b^2) without under/overflow}
\usage{hypot(a, b) }
\arguments{
  \item{a}{ value a}
  \item{b}{ value b}
}
\value{
  \item{double}{ sqrt(a^2 + b^2) }
}
\source{
Java Tutorial
\url{http://www.java2s.com/Tutorial/Java/0120__Development/sqrta2b2withoutunderoverflow.htm}
}
\references{
Java Tutorial
\url{http://www.java2s.com/Tutorial/Java/0120__Development/sqrta2b2withoutunderoverflow.htm}
}
\author{Jose Gama}
\examples{
hypot(0.310897, 0.306510)
}
\keyword{datasets}

