\name{LuminanceFactorToMunsellValue}
\alias{LuminanceFactorToMunsellValue}
\title{Convert luminance factor to Munsell value}
\description{\code{LuminanceFactorToMunsellValue} Converts the luminance
factor of an object colour, into a Munsell value
}
\usage{LuminanceFactorToMunsellValue(LumFac) }
\arguments{
  \item{LumFac}{ luminance factor  }
}
\value{ Munsell Value }
\source{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}
}
\references{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}
}
\author{Jose Gama}
\examples{
LuminanceFactorToMunsellValue(0.1)
}
\keyword{datasets}

