\name{MunsellRenotation}
\alias{MunsellRenotation}
\title{1943 Munsell renotation data}
\usage{MunsellRenotation}
\description{\code{MunsellRenotation} table with the 1943 Munsell renotation data (Munsell specifications in terms of CIE coordinates)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{H}{ Munsell hue}
    \item{V}{ Munsell value}
    \item{C}{ Munsell chroma}
    \item{x}{ CIE x}
    \item{y}{ CIE y}
    \item{Y}{ CIE Y}
  }
}
\source{
S. M. Newhall, D. Nickerson, & D. B. Judd 1943
"Final Report of the O.S.A. Subcommittee on the Spacing of the Munsell Colors"
Journal of the Optical Society of America, Vol. 33, Issue 7, pp. 385-418.
}
\references{
S. M. Newhall, D. Nickerson, & D. B. Judd 1943
"Final Report of the O.S.A. Subcommittee on the Spacing of the Munsell Colors"
Journal of the Optical Society of America, Vol. 33, Issue 7, pp. 385-418.
}
\author{Jose Gama}
\examples{
data(MunsellRenotation)
MunsellRenotation
}
\keyword{datasets}

