\name{MunsellValueToLuminanceFactorASTMD153508}
\alias{MunsellValueToLuminanceFactorASTMD153508}
\title{Convert Munsell value to luminance factor, ASTMD 1535-08}
\description{\code{MunsellValueToLuminanceFactorASTMD153508} Calculates the luminance
factor from the quintic polynomial in ASTMD 1535-08
}
\usage{MunsellValueToLuminanceFactorASTMD153508(MunsellValue) }
\arguments{
  \item{MunsellValue}{ Munsell value }
}
\value{ luminance factor }
\references{
ASTMD 1535-08

}
\author{Glenn Davis}
\examples{
MunsellValueToLuminanceFactorASTMD153508(5)
}
\keyword{datasets}

