\name{OptimalColoursForIlluminantC}
\alias{OptimalColoursForIlluminantC}
\alias{OptimalColoursForIlluminantD65}
\title{Table with optimal colors}
\usage{OptimalColoursForIlluminantC}
\description{\code{OptimalColoursForIlluminantC} table with optimal colors (on the boundary of the colour solid AKA MacAdam limits)}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{ CIE x}
    \item{y}{ CIE y}
    \item{Y}{ CIE Y}
  }
}
\source{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}

James D. Foley, Andries van Dam, Steven K. Feiner, & John F. Hughes, 1990
Computer Graphics: Principles and Practice, 2nd ed.,
Addison-Wesley Publishing Company.

Gunter Wyszecki & W. S. Stiles, 1982
Color Science: Concepts and Methods, Quantitative Data and Formulae, 2nd edition, John Wiley and Sons
}
\references{
Paul Centore 2014
The Munsell and Kubelka-Munk Toolbox
\url{http://www.99main.com/~centore/MunsellAndKubelkaMunkToolbox/MunsellAndKubelkaMunkToolbox.html}

James D. Foley, Andries van Dam, Steven K. Feiner, & John F. Hughes, 1990
Computer Graphics: Principles and Practice, 2nd ed.,
Addison-Wesley Publishing Company.

Gunter Wyszecki & W. S. Stiles, 1982
Color Science: Concepts and Methods, Quantitative Data and Formulae, 2nd edition, John Wiley and Sons
}
\author{Jose Gama}
\examples{
data(OptimalColoursForIlluminantC)
OptimalColoursForIlluminantC
}
\keyword{datasets}

