\name{BoundingRenotationHues}
\alias{BoundingRenotationHues}
\alias{CIELABtoApproxMunsellSpec}
\alias{CLHueLetterIndex}
\alias{CLHueNumPrefix}
\alias{ChromDiagHueAngleToMunsellHue}
\alias{ColorLabFormatToMunsellSpec}
\alias{FindHueOnRenotationOvoid}
\alias{LinearVsRadialInterpOnRenotationOvoid}
\alias{LuminanceFactorToMunsellValue}
\alias{MaxChromaForExtrapolatedRenotation}
\alias{MunsellHueToChromDiagHueAngle}
\alias{MunsellToxyForIntegerMunsellValue}
\alias{MunsellToxyYfromExtrapolatedRenotation}
\alias{getMunsellHueLetterDesignator}
\alias{MunsellValueToLuminanceFactor}
\alias{MunsellHueToASTMHue}
\alias{MunsellSpecToColorLabFormat}
\alias{lab2perc}
\alias{XYZ2xyY}
\alias{srgb2xyz}
\alias{xyY2XYZ}
\alias{xyz2lab}
\alias{xyz2srgb}
\alias{IsWithinMacAdamLimits}
\title{Internal RAHRS functions}
\description{Ignore these.}
\keyword{internal}

