\name{BoundingRenotationHues}
\alias{BoundingRenotationHues}
\alias{CIELABtoApproxMunsellSpec}
\alias{CLHueLetterIndex}
\alias{CLHueNumPrefix}
\alias{ChromDiagHueAngleToMunsellHue}
\alias{ColorLabFormatToMunsellSpec}
\alias{FindHueOnRenotationOvoid}
\alias{LinearVsRadialInterpOnRenotationOvoid}
\alias{MaxChromaForExtrapolatedRenotation}
\alias{MunsellHueToChromDiagHueAngle}
\alias{MunsellToxyForIntegerMunsellValue}
\alias{MunsellToxyYfromExtrapolatedRenotation}
\alias{getMunsellHueLetterDesignator}
\alias{MunsellHueToASTMHue}
\alias{MunsellSpecToColorLabFormat}
\alias{lab2perc}
\title{Internal munsellinterpolation functions}
\description{Ignore these.}
\keyword{internal}

