\name{xyY2XYZ}
\alias{xyY2XYZ}
\title{Convert xyY coordinates to XYZ}
\description{\code{xyY2XYZ} Converts xyY coordinates to XYZ.
}
\usage{xyY2XYZ(xyY) }
\arguments{
  \item{xyY}{ CIE xyY coordinates}
}
\value{
  CIE XYZ coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
xyY2XYZ(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

