\name{xyz2srgb}
\alias{xyz2srgb}
\title{Convert XYZ coordinates to sRGB}
\description{\code{xyz2srgb} Converts XYZ coordinates to sRGB.
}
\usage{xyz2srgb(XYZ) }
\arguments{
  \item{XYZ}{ XYZ coordinates}
}
\value{
  sRGB coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
xyz2srgb(c(0.310897, 0.306510, 74.613450))
}
\keyword{datasets}

