\name{XYZ2xyY}
\alias{XYZ2xyY}
\title{Convert XYZ coordinates to xyY}
\description{\code{XYZ2xyY} Converts XYZ coordinates to xyY.
}
\usage{XYZ2xyY(XYZ) }
\arguments{
  \item{XYZ}{ CIE XYZ coordinates}
}
\value{
  CIE xyY coordinates
}
\source{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\references{
Logicol S.r.l., 2014
EasyRGB color search engine
\url{http://www.easyrgb.com/}
}
\author{Jose Gama}
\examples{
XYZ2xyY(c(0.310897, 0.306510, 0.3))
}
\keyword{datasets}

